requirements.txt
scikit-learn
matplotlib
pandas
google.generativeai
certifi==2024.2.2
Flask==3.0.2
Flask-Cors==4.0.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
imutils==0.5.4
keras==2.15.0
matplotlib==3.8.2
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.0
pillow==10.2.0
progressbar==2.5
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
tensorflow==2.15.0
tqdm==4.66.2